fix: rename aws configuration file
