+++ packages/common/src/interfaces/IS3Config.ts
