+++ packages/common/src/interfaces/index.ts
@@ -5,9 +5,9 @@ export * from './IGraphql';
 export * from './IPlugin';
 export * from './IConfig';
 
-export * from './IAwsConfig';
-export * from './IWasabiConfig';
+export * from './IAppIntegrationConfig';
 export * from './IAuth0Config';
+export * from './IS3Config';
 export * from './IFacebookConfig';
 export * from './IFiverrConfig';
 export * from './IGithubConfig';
@@ -17,6 +17,6 @@ export * from './ILinkedinIConfig';
 export * from './IMicrosoftConfig';
 export * from './ISMTPConfig';
 export * from './ITwitterConfig';
-export * from './IUpworkConfig';
 export * from './IUnleashConfig';
-export * from './IAppIntegrationConfig';
+export * from './IUpworkConfig';
+export * from './IWasabiConfig';
